Unify object(fields) and variable definition around `wi_push_X`
This is more of a step toward adding arrays and maps to the API. I HAD to come up with a way to set object fields without a separate `wi_object_set_X`; same with variable definition: `wi_def_X` was pretty... limited. Now the API has `wi_def`, which is generic and just defines whatever is at the stack top. `wi_find_function` was kinda the same problem, now it's just `wi_find`, generic just like `wi_def`!

- Remove `wi_def_foreign`; replace with `wi_push_foreign` + `wi_def`
- Remove `wi_def_object`; replace with `wi_push_object` + `wi_def`
- Remove `wi_object_set_X`; replace with one `wi_object_set` that consumes the stack top
- Replace `wi_find_function` with `wi_find`
- Add `wi_foreign_entry` + `WI_DEF_FOREIGN_ALL`/`WI_OBJECT_SET_FOREIGN_ALL` for easier definition (replaces tons of `wi_push_foreign` + `wi_def`/`wi_object_set` calls)
- `foreign->arity`/`wi_new_foreign`/`wi_push_foreign` now take `uint8_t` instead of `int`, matching `wi_prototype`

+++ src/core/wi_box.c
@@ -129,7 +129,7 @@ wi_prototype_instr_size(struct wi_prototype* prototype, int offset) {
 }
 
 struct wi_foreign*
-wi_new_foreign(struct wi_gc* gc, wi_foreign_fn fn, int arity, bool is_variadic) {
+wi_new_foreign(struct wi_gc* gc, wi_foreign_fn fn, uint8_t arity, bool is_variadic) {
     struct wi_foreign* foreign = WI_NEW_BOX(gc, struct wi_foreign, WI_BOX_FOREIGN);
 
     foreign->fn          = fn;
